+++ RustApp/src/audio/mod.rs
@@ -9,7 +9,9 @@ use rtrb::{
     Consumer,
 };
 
-use crate::{config::AudioFormat, map_bytes::MapBytes, ui::app::AppState};
+pub mod player;
+
+use crate::{config::AudioFormat, ui::app::AppState};
 
 impl AppState {
     pub fn start_audio_stream(&self, consumer: Consumer<u8>) -> anyhow::Result<cpal::Stream> {
@@ -227,25 +229,81 @@ impl ChannelStrategy {
     }
 }
 
-fn print_supported_config(device: &Device) {
-    let output_configs = match device.supported_output_configs() {
-        Ok(f) => f.collect(),
-        Err(e) => {
-            println!("Error getting supported output configs: {:?}", e);
-            Vec::new()
-        }
-    };
-    if !output_configs.is_empty() {
-        println!("Supported configs:");
-        for conf in output_configs.into_iter() {
-            println!(
-                "channel:{}, min sample rate:{}, max sample rate:{}, audio format:{}",
-                conf.channels(),
-                conf.min_sample_rate().0,
-                conf.max_sample_rate().0,
-                conf.sample_format()
-            );
+pub trait MapBytes {
+    fn map_bytes<B>(iter: &mut impl Iterator<Item = u8>, sample_size: usize) -> Option<Self>
+    where
+        Self: Sized,
+        B: ByteOrder;
+}
+
+impl MapBytes for i16 {
+    fn map_bytes<B>(iter: &mut impl Iterator<Item = u8>, sample_size: usize) -> Option<Self>
+    where
+        Self: Sized,
+        B: ByteOrder,
+    {
+        if sample_size == 2 {
+            Some(B::read_i16(&[iter.next()?, iter.next()?]))
+        } else {
+            None
         }
     }
-    println!();
+}
+
+impl MapBytes for i32 {
+    fn map_bytes<B>(iter: &mut impl Iterator<Item = u8>, sample_size: usize) -> Option<Self>
+    where
+        Self: Sized,
+        B: ByteOrder,
+    {
+        if sample_size == 4 {
+            Some(B::read_i32(&[
+                iter.next()?,
+                iter.next()?,
+                iter.next()?,
+                iter.next()?,
+            ]))
+        } else {
+            None
+        }
+    }
+}
+
+impl MapBytes for f32 {
+    fn map_bytes<B>(iter: &mut impl Iterator<Item = u8>, sample_size: usize) -> Option<Self>
+    where
+        Self: Sized,
+        B: ByteOrder,
+    {
+        if sample_size == 3 {
+            // read 24 bits
+            let val = B::read_i24(&[iter.next()?, iter.next()?, iter.next()?]);
+
+            // convert to f32
+            Some(val as f32 / (1 << 23) as f32)
+        } else if sample_size == 4 {
+            Some(B::read_f32(&[
+                iter.next()?,
+                iter.next()?,
+                iter.next()?,
+                iter.next()?,
+            ]))
+        } else {
+            None
+        }
+    }
+}
+
+impl MapBytes for u8 {
+    fn map_bytes<B>(iter: &mut impl Iterator<Item = u8>, sample_size: usize) -> Option<Self>
+    where
+        Self: Sized,
+        B: ByteOrder,
+    {
+        if sample_size == 1 {
+            Some(iter.next()?)
+        } else {
+            None
+        }
+    }
 }
